Don't allow selecting text when using selection tool and support resizing when multiple elements include ones with binded text selectec
